--- packages/user-interface/src/tests/mock/data/AccountPageRequests.mock.tsx
@@ -1,3 +1,0 @@
-import { getBurgerProfiel, getPersoon } from "./AllRequests.mock";
-
-export const mocksRequestBurgerGegevens = [getPersoon, getBurgerProfiel];
