--- packages/user-interface/src/tests/mock/data/CasesPageRequests.mock.tsx
@@ -1,3 +1,0 @@
-import { getZaken } from "./AllRequests.mock";
-
-export const mockRequest = [getZaken];
