--- packages/user-interface/src/tests/mock/data/TasksPageRequests.mock.tsx
@@ -1,3 +1,0 @@
-import { getTaken } from "./AllRequests.mock";
-
-export const mockRequest = [getTaken()];
