+++ packages/user-interface/src/tests/mock/pages/CasesPage.mock.tsx
@@ -2,7 +2,7 @@ import CasesPage from "../../../pages/CasesPage";
 import TestProvider, {
   testPaths as paths,
 } from "../../../providers/TestProvider";
-import { mockRequest } from "../data/CasesPageRequests.mock";
+import { getZaken } from "../data/zaken.mock";
 
 const routes = [
   {
@@ -18,7 +18,10 @@ const routes = [
 
 export const MockCasesPage = () => (
   <TestProvider
-    mocks={mockRequest}
+    mocks={[
+      getZaken({ pageSize: 10, isOpen: true }),
+      getZaken({ pageSize: 10, isOpen: false }),
+    ]}
     routes={routes}
     initialIndex={0}
     initialEntries={[paths.cases]}
